move validateListOptions fun into `list` package
Move the func to the package it really belongs to.
