+++ backend/src/apiserver/server/list_request_util_test.go
@@ -17,15 +17,10 @@ package server
 import (
 	"encoding/base64"
 	"encoding/json"
-	"github.com/google/go-cmp/cmp/cmpopts"
-	"google.golang.org/protobuf/testing/protocmp"
 	"testing"
-
-	"github.com/google/go-cmp/cmp"
 
 	api "github.com/kubeflow/pipelines/backend/api/go_client"
 	"github.com/kubeflow/pipelines/backend/src/apiserver/common"
-	"github.com/kubeflow/pipelines/backend/src/apiserver/list"
 	"github.com/kubeflow/pipelines/backend/src/common/util"
 	"github.com/stretchr/testify/assert"
 	"google.golang.org/grpc/codes"
@@ -181,44 +176,6 @@ func TestParseSortByQueryString_StringTooLong(t *testing.T) {
 	assert.Contains(t, err.Error(), "Received invalid sort by format `Name desc foo`")
 }
 
-func TestParseAPIFilter_EmptyStringYieldsNilFilter(t *testing.T) {
-	f, err := parseAPIFilter("")
-
-	assert.Nil(t, err)
-	assert.Nil(t, f)
-}
-
-func TestParseAPIFilter_InvalidStringYieldsError(t *testing.T) {
-	f, err := parseAPIFilter("lkjlkjlkj")
-
-	assert.NotNil(t, err)
-	assert.Nil(t, f)
-}
-
-func TestParseAPIFilter_DecodesEncodedString(t *testing.T) {
-	// in was generated by calling JSON.stringify followed by encodeURIComponent in
-	// the browser on the following JSON string:
-	//   {"predicates":[{"op":"EQUALS","key":"testkey","stringValue":"testvalue"}]}
-
-	in := "%7B%22predicates%22%3A%5B%7B%22op%22%3A%22EQUALS%22%2C%22key%22%3A%22testkey%22%2C%22stringValue%22%3A%22testvalue%22%7D%5D%7D"
-
-	// The above should correspond the following filter:
-	want := &api.Filter{
-		Predicates: []*api.Predicate{
-			&api.Predicate{
-				Key: "testkey", Op: api.Predicate_EQUALS,
-				Value: &api.Predicate_StringValue{StringValue: "testvalue"},
-			},
-		},
-	}
-
-	got, err := parseAPIFilter(in)
-	if !cmp.Equal(got, want, cmpopts.EquateEmpty(), protocmp.Transform()) || err != nil {
-		t.Errorf("parseAPIString(%q) =\nGot %+v, %v\n Want %+v, <nil>\nDiff: %s",
-			in, got, err, want, cmp.Diff(want, got))
-	}
-}
-
 type fakeListable struct {
 	PrimaryKey       string
 	FakeName         string
@@ -273,26 +230,4 @@ func (f *fakeListable) GetSortByFieldPrefix(name string) string {
 
 func (f *fakeListable) GetKeyFieldPrefix() string {
 	return ""
-}
-
-func TestValidatedListOptions_Errors(t *testing.T) {
-	opts, err := list.NewOptions(&fakeListable{}, 10, "name asc", nil)
-	if err != nil {
-		t.Fatalf("list.NewOptions() = _, %+v; Want nil error", err)
-	}
-
-	npt, err := opts.NextPageToken(&fakeListable{})
-	if err != nil {
-		t.Fatalf("opt.NextPageToken() = _, %+v; Want nil error", err)
-	}
-
-	_, err = ValidatedListOptions(&fakeListable{}, npt, 10, "name asc", "")
-	if err != nil {
-		t.Fatalf("ValidatedListOptions(fakeListable, 10, \"name asc\") = _, %+v; Want nil error", err)
-	}
-
-	_, err = ValidatedListOptions(&fakeListable{}, npt, 10, "name desc", "")
-	if err == nil {
-		t.Fatalf("ValidatedListOptions(fakeListable, 10, \"name desc\") = _, %+v; Want error", err)
-	}
 }
